chore: moved SystemBroadcastReceiver.kt to its own component.
